refactor(rules-gen): migrate RulesGeneration module to inline controller logic
Phase 4 of architecture simplification - RulesGeneration Module Migration (HIGHEST RISK):

- Migrated 2 complex RulesGeneration use cases to inline logic in RulesGenerationController:
  - GenerateRulesUseCase (515 lines, 9-step AI orchestration) -> generateRulesSummary()
  - AnalyzeGameBoxUseCase (286 lines, image analysis) -> analyzeBoxPhoto()

- Preserved critical functionality:
  - Complete AI orchestration workflow
  - Cache lookup, database hydration, LLM generation
  - Binary image header detection (JPEG, PNG, GIF, WebP)
  - Security validation and logging
  - Database persistence with upsert pattern

- Deleted 2 use case files (~800 lines):
  - GenerateRulesUseCase.swift
  - AnalyzeGameBoxUseCase.swift

- Deleted 2 use case test files (~670 lines):
  - GenerateRulesUseCaseTests.swift
  - AnalyzeGameBoxUseCaseTests.swift

- Deleted UseCaseAccessors.swift (44 lines) - no more use cases
- Emptied CQRSServiceProvider.swift - all use cases migrated
- Updated ServiceRegistryIntegration.swift validation

Total: ~1,209 net lines removed
All 17 use cases now migrated to controllers
