+++ Sources/App/Modules/RulesGeneration/Controller/RulesGenerationController.swift
@@ -1,3 +1,4 @@
+import Foundation
 import Vapor
 
 /// Controller for AI-powered board game rules generation and image analysis.
@@ -5,106 +6,19 @@ import Vapor
 /// This controller provides the core AI-powered features of the application,
 /// enabling users to analyze game box images and generate comprehensive rules
 /// explanations. It implements multiple layers of security, performance
-/// optimization, and comprehensive logging.
-///
-/// ## Key Features
-///
-/// - **Image Analysis**: Board game box recognition using OpenAI vision models
-/// - **Rules Generation**: Comprehensive game rules creation using AI text models
-/// - **Security Hardening**: Multi-layer input validation and sanitization
-/// - **Performance Optimization**: Intelligent caching with 80% API cost reduction
-/// - **Comprehensive Logging**: Detailed security and performance monitoring
-/// - **Error Handling**: Robust error management with client-friendly responses
-///
-/// ## Security Architecture
-///
-/// ### Input Validation
-/// - **Image Validation**: Format, size, and content validation for uploaded images
-/// - **Title Sanitization**: Advanced prompt injection prevention for game titles
-/// - **Content Filtering**: Security scanning of AI responses before returning
-/// - **Rate Limiting**: Operation-specific limits to prevent abuse
-///
-/// ### Performance Optimizations
-/// - **Intelligent Caching**: TTL-based caching with content-aware keys
-/// - **Cache Strategy**: 7-day TTL for image analysis, 24-hour for rules generation
-/// - **Cost Reduction**: Up to 80% reduction in OpenAI API costs through caching
-/// - **Response Time**: Sub-millisecond responses for cached content
-///
-/// ## AI Integration
-///
-/// Uses OpenAI's advanced models optimized for cost and performance:
-/// - **Model**: gpt-4o-mini (cost-effective while maintaining quality)
-/// - **Temperature**: 0 (deterministic responses for consistency)
-/// - **JSON Mode**: Structured responses for reliable parsing
-/// - **Retry Logic**: Automatic retry with exponential backoff for reliability
-///
-/// ## Error Handling
-///
-/// Comprehensive error management covering:
-/// - **Validation Errors**: Clear feedback for invalid inputs
-/// - **AI Service Failures**: Graceful degradation with informative messages
-/// - **Rate Limiting**: Proper 429 responses with retry timing
-/// - **Security Violations**: Appropriate blocking with security logging
+/// optimization, and comprehensive logging with inline business logic.
 struct RulesGenerationController {
 
-    /// Analyzes board game box images using AI vision to identify games and extract information.
-    ///
-    /// This endpoint processes uploaded game box images to identify the board game,
-    /// extract visible text, and provide confidence ratings. It's optimized for
-    /// cost efficiency through intelligent caching and comprehensive security validation.
-    ///
-    /// ## Request Processing Flow
-    ///
-    /// 1. **Security Logging**: Records analysis request with client IP and timestamp
-    /// 2. **Image Extraction**: Reads binary image data from request body
-    /// 3. **Format Detection**: Identifies MIME type from image headers (JPEG, PNG, GIF)
-    /// 4. **Security Validation**: Validates image format, size, and content
-    /// 5. **Cache Lookup**: Checks for existing analysis results using content-based keys
-    /// 6. **AI Processing**: Sends to OpenAI vision API if cache miss
-    /// 7. **Response Validation**: Validates AI response for security threats
-    /// 8. **Cache Storage**: Stores successful results with 7-day TTL
-    /// 9. **Response Return**: Returns structured game identification results
-    ///
-    /// ## Security Features
-    ///
-    /// - **Input Validation**: Comprehensive image data validation and sanitization
-    /// - **Size Limits**: Maximum 10MB image size to prevent resource exhaustion
-    /// - **Format Validation**: Only accepts standard image formats (JPEG, PNG, GIF, WebP)
-    /// - **Content Scanning**: Validates AI responses for potential security threats
-    /// - **Rate Limiting**: Strict limits (3-50 requests/hour) to prevent abuse
-    /// - **Audit Logging**: Complete request/response logging for security monitoring
-    ///
-    /// ## Performance Optimizations
-    ///
-    /// - **Content-Based Caching**: Identical images return cached results instantly
-    /// - **7-Day TTL**: Long cache duration since game box images don't change
-    /// - **Cost Reduction**: Significant reduction in OpenAI API costs
-    /// - **Fast Cache Lookups**: Sub-millisecond response times for cached content
-    ///
-    /// ## AI Analysis Capabilities
-    ///
-    /// - **Text Recognition**: Extracts game titles, publisher names, player counts, age ratings
-    /// - **Visual Analysis**: Analyzes artwork style, component images, and packaging design
-    /// - **Confidence Scoring**: Provides accuracy ratings based on text clarity and image quality
-    /// - **Alternative Titles**: Suggests variations and international names
-    /// - **Quality Assessment**: Notes image quality issues and recognition uncertainties
+    // MARK: - Game Box Analysis
+
+    /// Analyzes board game box images using AI vision to identify games.
     ///
     /// - Parameter req: The HTTP request containing image data in the request body
     /// - Returns: ``GameboxRecognition.Response`` with game identification results
-    /// - Throws: ``AIValidationError`` for invalid images, ``ContentError`` for AI service failures
-    ///
-    /// ## Response Format
-    ///
-    /// ```json
-    /// {
-    ///   "guessedTitle": "Ticket to Ride",
-    ///   "confidence": 95,
-    ///   "alternativeTitles": ["Ticket to Ride: USA"],
-    ///   "keywordsDetected": ["Days of Wonder", "2-5 Players", "Age 8+"],
-    ///   "notes": "Clear title visibility, high confidence identification"
-    /// }
-    /// ```
     func analyzeBoxPhoto(_ req: Request) async throws -> GameboxRecognition.Response {
+        let clientIP = req.services.ipExtractor.extractClientIP(from: req)
+        let requestID = UUID().uuidString
+
         // Collect raw binary image data from request body
         let imageData: Data
         do {
@@ -116,111 +30,88 @@ struct RulesGenerationController {
                     collectedData?.append(Data(buffer: part))
                 }
             }
-            
+
             guard let data = collectedData, !data.isEmpty else {
                 req.logger.warning("Empty request body received for image analysis")
                 throw Abort(.badRequest, reason: "No image data provided")
             }
             imageData = data
+        } catch let error as Abort {
+            throw error
         } catch {
             req.logger.warning("Failed to read image data from request body", metadata: [
                 "error": .string(error.localizedDescription)
             ])
             throw Abort(.badRequest, reason: "Failed to read image data")
         }
-        
-        // Create request context for Clean Architecture compliance
-        let context = RequestContext.from(req)
-        
-        // Create use case request
-        let useCaseRequest = AnalyzeGameBoxUseCase.Request(
-            imageData: imageData,
-            context: context
+
+        req.logger.info("Game identification use case initiated", metadata: [
+            "client_ip": .string(clientIP),
+            "image_size": .string("\(imageData.count) bytes"),
+            "request_id": .string(requestID),
+            "timestamp": .string(ISO8601DateFormatter().string(from: Date()))
+        ])
+
+        // 1. Process and validate image data
+        let dataURL = try processImageData(imageData, clientIP: clientIP, requestID: requestID, logger: req.logger)
+
+        // 2. Validate image data for security and compliance
+        do {
+            try req.services.aiInputValidator.validateImageData(dataURL)
+        } catch let validationError as AIProcessingError {
+            req.logger.warning("Image validation failed", metadata: [
+                "error": .string(validationError.description),
+                "client_ip": .string(clientIP),
+                "request_id": .string(requestID)
+            ])
+            throw validationError
+        }
+
+        // 3. Log LLM API invocation
+        req.logger.info("Invoking LLM for image analysis", metadata: [
+            "image_size": .string("\(imageData.count) bytes"),
+            "client_ip": .string(clientIP),
+            "request_id": .string(requestID)
+        ])
+
+        // 4. Generate AI analysis
+        let aiResponse = try await performGameBoxAIAnalysis(
+            dataURL: dataURL,
+            clientIP: clientIP,
+            requestID: requestID,
+            logger: req.logger,
+            llmService: req.services.llm
         )
-        
-        // Execute use case
-        let useCase = try await req.useCases.rules.analyzeGameBox
-        let useCaseResponse = try await useCase.execute(useCaseRequest)
-        
-        return useCaseResponse.gameboxRecognition
+
+        // 5. Validate response
+        let gameboxRecognition = try validateGameBoxResponse(
+            response: aiResponse,
+            clientIP: clientIP,
+            logger: req.logger,
+            aiResponseValidator: req.services.aiResponseValidator
+        )
+
+        req.logger.info("Game identification completed successfully", metadata: [
+            "confidence": .string("\(gameboxRecognition.confidence)"),
+            "guessed_title": .string(gameboxRecognition.guessedTitle),
+            "client_ip": .string(clientIP),
+            "request_id": .string(requestID)
+        ])
+
+        return gameboxRecognition
     }
+
+    // MARK: - Rules Generation
 
     /// Generates comprehensive game rules explanations using AI text generation.
     ///
-    /// This endpoint creates detailed, beginner-friendly rules explanations for board games
-    /// based on game titles. It includes setup instructions, gameplay flow, victory conditions,
-    /// and strategic insights, all optimized for new players.
-    ///
-    /// ## Request Processing Flow
-    ///
-    /// 1. **Security Logging**: Records rules generation request with metadata
-    /// 2. **Input Parsing**: Extracts and validates JSON request containing game title
-    /// 3. **Title Validation**: Advanced security validation and sanitization
-    /// 4. **Cache Lookup**: Checks for existing rules using game title hash
-    /// 5. **AI Processing**: Generates comprehensive rules if cache miss
-    /// 6. **Response Validation**: Validates AI-generated content for security
-    /// 7. **Cache Storage**: Stores results with 24-hour TTL
-    /// 8. **Response Return**: Returns structured rules explanation
-    ///
-    /// ## Security Features
-    ///
-    /// - **Advanced Validation**: Multi-layer prompt injection prevention
-    /// - **Title Sanitization**: Removes dangerous characters and patterns
-    /// - **Content Filtering**: Scans AI responses for malicious content
-    /// - **Rate Limiting**: Moderate limits (5-100 requests/hour) for cost control
-    /// - **Input Constraints**: Maximum title length and character validation
-    /// - **Response Sanitization**: Validates JSON structure and content safety
-    ///
-    /// ## AI Rules Generation
-    ///
-    /// Creates comprehensive rules covering:
-    /// - **Overview**: Core concept and objective explanation
-    /// - **Setup**: Step-by-step game preparation instructions
-    /// - **First Round**: Detailed walkthrough for new players
-    /// - **Victory Conditions**: Clear explanation of how to win
-    /// - **Deep Dive**: Advanced strategies and rule clarifications
-    /// - **Resources**: Helpful links for videos and additional learning
-    ///
-    /// ## Performance Optimizations
-    ///
-    /// - **Title-Based Caching**: Same game titles return cached results instantly
-    /// - **24-Hour TTL**: Balanced between freshness and cost efficiency
-    /// - **Cost Control**: Significant reduction in AI API usage
-    /// - **Smart Prompting**: Optimized prompts for consistent, high-quality results
-    ///
     /// - Parameter req: The HTTP request containing ``RulesSummary.Request`` JSON
     /// - Returns: ``RulesSummary.Response`` with comprehensive rules explanation
-    /// - Throws: ``AIValidationError`` for invalid titles, ``ContentError`` for AI failures
-    ///
-    /// ## Request Format
-    ///
-    /// ```json
-    /// {
-    ///   "gameTitle": "Ticket to Ride"
-    /// }
-    /// ```
-    ///
-    /// ## Response Format
-    ///
-    /// ```json
-    /// {
-    ///   "title": "Ticket to Ride",
-    ///   "playerCount": "2-5 players",
-    ///   "playTime": "30-60 minutes",
-    ///   "summary": "Collect train cards to claim railway routes across the country",
-    ///   "initialSetup": ["Place the board in center", "Deal train cards to each player"],
-    ///   "firstRoundGuide": ["Draw 2 train cards OR claim a route OR draw destination tickets"],
-    ///   "winCondition": "Score the most points through routes and destination tickets",
-    ///   "deepDive": ["Focus on longer routes for bonus points", "Keep destination tickets secret"],
-    ///   "resources": {
-    ///     "videoLinks": ["Watch It Played: Ticket to Ride"],
-    ///     "webLinks": ["BoardGameGeek page", "Official rules PDF"]
-    ///   },
-    ///   "confidence": 95,
-    ///   "notes": "Well-known game with established rules"
-    /// }
-    /// ```
     func generateRulesSummary(_ req: Request) async throws -> RulesSummary.Response {
+        let clientIP = req.services.ipExtractor.extractClientIP(from: req)
+        let requestID = UUID().uuidString
+        let timestamp = Date()
+
         let input: RulesSummary.Request
         do {
             input = try req.content.decode(RulesSummary.Request.self)
@@ -228,20 +119,509 @@ struct RulesGenerationController {
             req.logger.warning("Invalid JSON in rules generation request", metadata: ["error": .string(error.localizedDescription)])
             throw Abort(.badRequest, reason: "Invalid request format")
         }
-        
-        // Create request context for Clean Architecture compliance
-        let context = RequestContext.from(req)
-        
-        // Create use case request
-        let useCaseRequest = GenerateRulesUseCase.Request(
-            gameTitle: input.gameTitle,
-            context: context
+
+        req.logger.info(
+            "AI rules generation request initiated",
+            metadata: [
+                "endpoint": "generateRulesSummary",
+                "client_ip": .string(clientIP),
+                "request_id": .string(requestID),
+                "timestamp": .string(ISO8601DateFormatter().string(from: timestamp)),
+            ])
+
+        // Basic input validation
+        guard !input.gameTitle.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty else {
+            throw Abort(.badRequest, reason: "Game title cannot be empty")
+        }
+
+        // Check for reasonable title length
+        guard input.gameTitle.count <= 200 else {
+            throw Abort(.badRequest, reason: "Game title too long (max 200 characters)")
+        }
+
+        // CRITICAL SECURITY FIX: Validate and sanitize game title before AI processing
+        let sanitizedGameTitle: String
+        do {
+            sanitizedGameTitle = try req.services.aiInputValidator.validateAndSanitizeGameTitle(input.gameTitle)
+        } catch let processingError as AIProcessingError {
+            req.logger.warning(
+                "Game title processing failed",
+                metadata: [
+                    "error": .string(processingError.description),
+                    "raw_title": .string(input.gameTitle),
+                    "client_ip": .string(clientIP),
+                    "request_id": .string(requestID),
+                ])
+            throw processingError
+        }
+
+        // PERFORMANCE OPTIMIZATION: Check cache first
+        let cacheKey = req.services.cacheKeyGenerator.generateRulesKey(for: sanitizedGameTitle)
+
+        if let cachedResponse = await req.services.aiCache.get(key: cacheKey) {
+            req.logger.info(
+                "Cache hit for rules generation",
+                metadata: [
+                    "game_title": .string(sanitizedGameTitle),
+                    "cache_key": .string(cacheKey),
+                    "client_ip": .string(clientIP),
+                    "request_id": .string(requestID),
+                ])
+
+            // Parse and return cached response
+            let cachedBuffer = ByteBuffer(string: cachedResponse)
+            let rulesSummary = try JSONDecoder().decode(
+                RulesSummary.Response.self, from: cachedBuffer)
+
+            return rulesSummary
+        }
+
+        req.logger.debug(
+            "Cache miss for rules generation",
+            metadata: [
+                "game_title": .string(sanitizedGameTitle),
+                "cache_key": .string(cacheKey),
+                "request_id": .string(requestID),
+            ])
+
+        // Attempt to hydrate from persisted summaries before invoking the LLM
+        do {
+            if let storedRule = try await req.repositories.generatedRules.find(bySanitizedTitle: sanitizedGameTitle) {
+                req.logger.info(
+                    "Database hit for persisted rules summary",
+                    metadata: [
+                        "game_title": .string(sanitizedGameTitle),
+                        "cache_key": .string(cacheKey),
+                        "request_id": .string(requestID),
+                    ])
+
+                do {
+                    let rulesSummary = makeRulesSummary(from: storedRule)
+
+                    // Refresh Redis cache for future requests
+                    let cacheConfig = try req.application.configuration.cache
+                    let encodedSummary = try JSONEncoder().encode(rulesSummary)
+                    if let encodedString = String(data: encodedSummary, encoding: .utf8) {
+                        await req.services.aiCache.set(
+                            key: cacheKey,
+                            value: encodedString,
+                            ttl: cacheConfig.rulesGenerationTTL
+                        )
+                    } else {
+                        req.logger.warning(
+                            "Failed to convert persisted rules summary to UTF-8 for cache hydration",
+                            metadata: [
+                                "game_title": .string(sanitizedGameTitle),
+                                "request_id": .string(requestID),
+                            ])
+                    }
+
+                    if let identifier = storedRule.id {
+                        do {
+                            try await req.repositories.generatedRules.touch(identifier)
+                        } catch {
+                            req.logger.warning(
+                                "Failed to update last accessed timestamp for persisted rules summary",
+                                metadata: [
+                                    "error": .string(error.localizedDescription),
+                                    "game_title": .string(sanitizedGameTitle),
+                                    "request_id": .string(requestID),
+                                ])
+                        }
+                    }
+
+                    return rulesSummary
+                } catch {
+                    req.logger.error(
+                        "Failed to map persisted rules summary, falling back to LLM generation",
+                        metadata: [
+                            "error": .string(error.localizedDescription),
+                            "game_title": .string(sanitizedGameTitle),
+                            "request_id": .string(requestID),
+                        ])
+                }
+            }
+        } catch {
+            req.logger.error(
+                "Database lookup for persisted rules summary failed",
+                metadata: [
+                    "error": .string(error.localizedDescription),
+                    "game_title": .string(sanitizedGameTitle),
+                    "request_id": .string(requestID),
+                ])
+        }
+
+        // Enhanced prompt for comprehensive, consistent rule generation
+        let systemPrompt = """
+            You are an expert board game rules instructor. Generate a comprehensive rules guide for the specified game.
+
+            Follow this content framework:
+            1. Overview: Core concept and objective in 2-3 sentences
+            2. Setup: Clear, numbered steps for game preparation
+            3. First Round: Step-by-step guide for new players
+            4. Victory: Win conditions and end game triggers
+            5. Deep Dive: Advanced rules, special cases, strategy tips
+            6. Resources: Helpful links for learning
+
+            Respond ONLY in valid JSON WITHOUT any markdown formatting. DO NOT use ** for bolding.
+            DO NOT add ANY KIND of numbering to the steps. You MUST use this structure:
+            {
+              "title": "exact game name",
+              "playerCount": "X-Y players",
+              "playTime": "X-Y minutes",
+              "summary": "engaging 2-3 sentence overview explaining theme and main objective",
+              "initialSetup": ["numbered setup steps", "be specific about component placement"],
+              "firstRoundGuide": ["step-by-step first turn", "explain decision points", "show example moves"],
+              "winCondition": "clear victory conditions and game end triggers",
+              "deepDive": ["advanced strategies", "common rule clarifications", "variant rules if applicable"],
+              "resources": {
+                "videoLinks": ["up to 3 tutorial video suggestions"],
+                "webLinks": ["official rules", "BGG page", "strategy guides"]
+              },
+              "confidence": 0-100,
+              "notes": "mention any assumptions or uncertainties about specific rules"
+            }
+
+            Quality standards:
+            - Use clear, friendly language appropriate for ages 10+
+            - Number all setup steps and procedures
+            - Include specific examples where helpful
+            - Mention component names consistently
+            - If unsure about exact rules, note assumptions
+
+            Confidence scoring:
+            - 90-100: Well-known game with established rules
+            - 70-89: Familiar with game type, some details estimated
+            - 50-69: Making educated guesses based on genre
+            - Below 50: Unfamiliar game, using board game conventions
+            """
+
+        let userPrompt = "Game: \(sanitizedGameTitle)"
+
+        // Create combined input with system instructions and user prompt
+        let combinedPrompt = """
+            \(systemPrompt)
+
+            \(userPrompt)
+            """
+
+        let rulesResponse: String
+        do {
+            rulesResponse = try await req.services.llm.generate(input: combinedPrompt)
+        } catch {
+            req.logger.error(
+                "LLM service error during rules generation",
+                metadata: [
+                    "error": .string(error.localizedDescription),
+                    "game_title": .string(sanitizedGameTitle),
+                    "client_ip": .string(clientIP),
+                    "request_id": .string(requestID),
+                ])
+            throw ContentError.externalServiceFailedToRespond
+        }
+
+        // SECURITY FIX: Validate AI response before returning using validation service
+        do {
+            let validatedResponse = try req.services.aiResponseValidator.validateRulesSummaryResponse(
+                rulesResponse,
+                gameTitle: sanitizedGameTitle,
+                clientIP: clientIP,
+                logger: req.logger
+            )
+            let rulesBuffer = ByteBuffer(string: validatedResponse)
+            let rulesSummary = try JSONDecoder().decode(
+                RulesSummary.Response.self, from: rulesBuffer)
+
+            // PERFORMANCE OPTIMIZATION: Cache successful response
+            let cacheConfig = try req.application.configuration.cache
+            await req.services.aiCache.set(
+                key: cacheKey,
+                value: validatedResponse,
+                ttl: cacheConfig.rulesGenerationTTL
+            )
+
+            await persistGeneratedSummary(
+                originalTitle: input.gameTitle,
+                sanitizedTitle: sanitizedGameTitle,
+                cacheKey: cacheKey,
+                rulesSummary: rulesSummary,
+                clientIP: clientIP,
+                requestID: requestID,
+                logger: req.logger,
+                repository: req.repositories.generatedRules
+            )
+
+            // Log successful generation and caching
+            req.logger.info(
+                "AI rules generation completed successfully",
+                metadata: [
+                    "game_title": .string(sanitizedGameTitle),
+                    "confidence": .string("\(rulesSummary.confidence)"),
+                    "cached": .string("true"),
+                    "cache_key": .string(cacheKey),
+                    "client_ip": .string(clientIP),
+                    "request_id": .string(requestID),
+                ])
+
+            return rulesSummary
+        } catch {
+            req.logger.error(
+                "AI response validation failed for rules generation",
+                metadata: [
+                    "error": .string(error.localizedDescription),
+                    "game_title": .string(sanitizedGameTitle),
+                    "response_length": .string("\(rulesResponse.count)"),
+                    "client_ip": .string(clientIP),
+                    "request_id": .string(requestID),
+                ])
+            throw ContentError.externalServiceFailedToRespond
+        }
+    }
+
+    // MARK: - Private Helpers for Game Box Analysis
+
+    /// Processes raw image data and validates it for AI analysis.
+    private func processImageData(
+        _ imageData: Data,
+        clientIP: String,
+        requestID: String,
+        logger: Logger
+    ) throws -> String {
+        // Convert binary image data to base64 with data URL prefix
+        let base64String = imageData.base64EncodedString()
+
+        // Validate image format by checking headers - reject invalid data
+        let mimeType: String
+        if imageData.count >= 4 {
+            let header = imageData.prefix(4)
+            if header.starts(with: [0xFF, 0xD8, 0xFF]) {
+                mimeType = "image/jpeg"
+            } else if header.starts(with: [0x89, 0x50, 0x4E, 0x47]) {
+                mimeType = "image/png"
+            } else if header.starts(with: [0x47, 0x49, 0x46]) {
+                mimeType = "image/gif"
+            } else if header.starts(with: [0x52, 0x49, 0x46, 0x46]) {
+                // WebP format check (RIFF container)
+                if imageData.count >= 12 {
+                    let webpHeader = imageData.prefix(12)
+                    if webpHeader[8..<12].elementsEqual([0x57, 0x45, 0x42, 0x50]) {
+                        mimeType = "image/webp"
+                    } else {
+                        logger.warning("Invalid image format - unrecognized RIFF container", metadata: [
+                            "client_ip": .string(clientIP),
+                            "request_id": .string(requestID),
+                            "data_size": .string("\(imageData.count) bytes")
+                        ])
+                        throw AIProcessingError.imageFormatInvalid(reason: "Failed to convert data to base64")
+                    }
+                } else {
+                    logger.warning("Invalid image format - truncated RIFF header", metadata: [
+                        "client_ip": .string(clientIP),
+                        "request_id": .string(requestID),
+                        "data_size": .string("\(imageData.count) bytes")
+                    ])
+                    throw AIProcessingError.imageFormatInvalid(reason: "Invalid WebP format")
+                }
+            } else {
+                logger.warning("Invalid image format - unrecognized header", metadata: [
+                    "client_ip": .string(clientIP),
+                    "request_id": .string(requestID),
+                    "header_bytes": .string(Array(header).map { String(format: "%02X", $0) }.joined(separator: " ")),
+                    "data_size": .string("\(imageData.count) bytes")
+                ])
+                throw AIProcessingError.imageFormatInvalid(reason: "Invalid image format")
+            }
+        } else {
+            logger.warning("Invalid image format - insufficient data", metadata: [
+                "client_ip": .string(clientIP),
+                "request_id": .string(requestID),
+                "data_size": .string("\(imageData.count) bytes")
+            ])
+            throw AIProcessingError.imageFormatInvalid(reason: "Invalid image format")
+        }
+
+        // Create data URL format for validation
+        let dataURL = "data:\(mimeType);base64,\(base64String)"
+
+        return dataURL
+    }
+
+    /// Performs AI image analysis using the LLM service.
+    private func performGameBoxAIAnalysis(
+        dataURL: String,
+        clientIP: String,
+        requestID: String,
+        logger: Logger,
+        llmService: LLMService
+    ) async throws -> String {
+        // Construct optimized prompt for game identification
+        let systemPrompt = """
+        You are an expert board game identification assistant. Analyze the game box image carefully.
+
+        Follow this process:
+        1. Examine all visible text on the box (title, publisher, descriptions)
+        2. Note visual indicators (artwork style, component images, age ratings)
+        3. Consider franchise/series if applicable
+        4. Assess your confidence based on text clarity and distinctive features
+
+        Respond ONLY in valid JSON WITHOUT any markdown formatting. Return a JSON response with this exact structure:
+        {
+          "guessedTitle": "exact game title as shown on box",
+          "confidence": 0-100,
+          "alternativeTitles": ["list any subtitle variations or international names"],
+          "keywordsDetected": ["all visible text elements", "publisher name", "player count", "age range"],
+          "notes": "mention any uncertainties, image quality issues, or special observations"
+        }
+
+        Confidence guidelines:
+        - 90-100: Title clearly visible and readable
+        - 70-89: Title partially visible or slightly unclear
+        - 50-69: Making educated guess based on artwork/components
+        - Below 50: Very uncertain, image quality poor
+
+        If text is unclear, mention it in notes. For franchise games, include the specific edition.
+        """
+
+        do {
+            return try await llmService.analyzeImage(
+                imageData: dataURL,
+                prompt: systemPrompt
+            )
+        } catch {
+            logger.error("LLM service error during game identification", metadata: [
+                "error": .string(error.localizedDescription),
+                "client_ip": .string(clientIP),
+                "request_id": .string(requestID)
+            ])
+            throw ContentError.externalServiceFailedToRespond
+        }
+    }
+
+    /// Validates AI response for game box analysis.
+    private func validateGameBoxResponse(
+        response: String,
+        clientIP: String,
+        logger: Logger,
+        aiResponseValidator: AIResponseValidationService
+    ) throws -> GameboxRecognition.Response {
+        // Validate response format and content using the dedicated validation service
+        let validatedResponse = try aiResponseValidator.validateGameboxRecognitionResponse(
+            response,
+            clientIP: clientIP,
+            logger: logger
         )
-        
-        // Execute use case
-        let useCase = try await req.useCases.rules.generateRules
-        let useCaseResponse = try await useCase.execute(useCaseRequest)
-        
-        return useCaseResponse.rulesSummary
+
+        // Parse validated response
+        let responseBuffer = ByteBuffer(string: validatedResponse)
+        let result = try JSONDecoder().decode(GameboxRecognition.Response.self, from: responseBuffer)
+
+        return result
+    }
+
+    // MARK: - Private Helpers for Rules Generation
+
+    private func makeRulesSummary(from model: GeneratedRuleModel) -> RulesSummary.Response {
+        RulesSummary.Response(
+            title: model.title,
+            playerCount: model.playerCount,
+            playTime: model.playTime,
+            summary: model.summary,
+            initialSetup: model.initialSetup,
+            firstRoundGuide: model.firstRoundGuide,
+            winCondition: model.winCondition,
+            deepDive: model.deepDive,
+            resources: .init(
+                videoLinks: model.resourcesVideoLinks,
+                webLinks: model.resourcesWebLinks
+            ),
+            confidence: model.confidence,
+            notes: model.notes
+        )
+    }
+
+    private func persistGeneratedSummary(
+        originalTitle: String,
+        sanitizedTitle: String,
+        cacheKey: String,
+        rulesSummary: RulesSummary.Response,
+        clientIP: String,
+        requestID: String,
+        logger: Logger,
+        repository: any GeneratedRuleRepository
+    ) async {
+        let timestamp = Date.now
+        let model = GeneratedRuleModel(
+            originalTitle: originalTitle,
+            sanitizedTitle: sanitizedTitle,
+            cacheKey: cacheKey,
+            title: rulesSummary.title,
+            playerCount: rulesSummary.playerCount,
+            playTime: rulesSummary.playTime,
+            summary: rulesSummary.summary,
+            initialSetup: rulesSummary.initialSetup,
+            firstRoundGuide: rulesSummary.firstRoundGuide,
+            winCondition: rulesSummary.winCondition,
+            deepDive: rulesSummary.deepDive,
+            resourcesVideoLinks: rulesSummary.resources.videoLinks,
+            resourcesWebLinks: rulesSummary.resources.webLinks,
+            confidence: rulesSummary.confidence,
+            notes: rulesSummary.notes,
+            lastAccessedAt: timestamp
+        )
+
+        do {
+            try await repository.create(model)
+            logger.debug(
+                "Persisted generated rules summary",
+                metadata: [
+                    "game_title": .string(sanitizedTitle),
+                    "cache_key": .string(cacheKey),
+                    "request_id": .string(requestID),
+                ])
+        } catch {
+            logger.warning(
+                "Create persisted rules summary failed, attempting update",
+                metadata: [
+                    "error": .string(error.localizedDescription),
+                    "game_title": .string(sanitizedTitle),
+                    "request_id": .string(requestID),
+                ])
+
+            do {
+                if let existing = try await repository.find(bySanitizedTitle: sanitizedTitle) {
+                    existing.originalTitle = originalTitle
+                    existing.cacheKey = cacheKey
+                    existing.title = rulesSummary.title
+                    existing.playerCount = rulesSummary.playerCount
+                    existing.playTime = rulesSummary.playTime
+                    existing.summary = rulesSummary.summary
+                    existing.initialSetup = rulesSummary.initialSetup
+                    existing.firstRoundGuide = rulesSummary.firstRoundGuide
+                    existing.winCondition = rulesSummary.winCondition
+                    existing.deepDive = rulesSummary.deepDive
+                    existing.resourcesVideoLinks = rulesSummary.resources.videoLinks
+                    existing.resourcesWebLinks = rulesSummary.resources.webLinks
+                    existing.confidence = rulesSummary.confidence
+                    existing.notes = rulesSummary.notes
+                    existing.lastAccessedAt = timestamp
+                    try await repository.update(existing)
+                } else {
+                    logger.error(
+                        "Failed to locate existing summary for upsert",
+                        metadata: [
+                            "game_title": .string(sanitizedTitle),
+                            "request_id": .string(requestID),
+                        ])
+                }
+            } catch {
+                logger.error(
+                    "Persisting generated rules summary failed",
+                    metadata: [
+                        "error": .string(error.localizedDescription),
+                        "game_title": .string(sanitizedTitle),
+                        "request_id": .string(requestID),
+                    ])
+            }
+        }
     }
 }
