--- Tests/AppTests/UseCases/RulesGeneration/GenerateRulesUseCaseTests.swift
@@ -1,384 +1,0 @@
-import Testing
-import Vapor
-@testable import App
-
-/// Comprehensive tests for GenerateRulesUseCase demonstrating Complex Command testing patterns.
-///
-/// This test suite validates AI-powered rules generation business logic in isolation,
-/// including orchestration of multiple AI services, caching strategies, and complex
-/// error handling scenarios for multi-step AI operations.
-@Suite(.serialized)
-final class GenerateRulesUseCaseTests: Sendable {
-    let world: IsolatedTestWorld
-    let app: Application
-    let cache: MockAICacheService
-    let repository: TestGeneratedRuleRepository
-
-    init() async throws {
-        world = try await IsolatedTestWorld()
-        app = world.app
-        cache = world.aiCache
-        repository = world.generatedRules
-        try await app.autoMigrate()
-        await world.resetAll()
-    }
-
-    /// Test use case request structure and validation.
-    @Test("GenerateRulesUseCase Request has correct structure")
-    func testRequestStructure() async throws {
-        await world.resetAll()
-        // Arrange
-        let context = RequestContext(
-            clientIP: "127.0.0.1",
-            logger: Logger(label: "test"),
-            timestamp: Date(),
-            requestID: "test-123"
-        )
-        
-        // Act
-        let request = GenerateRulesUseCase.Request(
-            gameTitle: "Ticket to Ride",
-            context: context
-        )
-        
-        // Assert
-        #expect(request.gameTitle == "Ticket to Ride")
-        #expect(request.context.clientIP == "127.0.0.1")
-        #expect(request.context.requestID == "test-123")
-    }
-    
-    /// Test use case response structure and validation.
-    @Test("GenerateRulesUseCase Response has correct structure")
-    func testResponseStructure() async throws {
-        await world.resetAll()
-        // Arrange
-        let rulesSummary = RulesSummary.Response(
-            title: "Monopoly Rules",
-            playerCount: "2-8",
-            playTime: "60-180 minutes",
-            summary: "A classic property trading game",
-            initialSetup: ["Set up the board", "Distribute money"],
-            firstRoundGuide: ["Roll dice", "Move token", "Buy property"],
-            winCondition: "Be the last player with money",
-            deepDive: ["Property development strategy", "Cash flow management"],
-            resources: RulesSummary.Response.GameResources(
-                videoLinks: ["https://youtube.com/monopoly"],
-                webLinks: ["https://monopoly.com/rules"]
-            ),
-            confidence: 90,
-            notes: "Complete official rules"
-        )
-        
-        // Act
-        let response = GenerateRulesUseCase.Response(
-            rulesSummary: rulesSummary,
-            processedGameTitle: "Monopoly",
-            generatedAt: Date(),
-            wasCached: false
-        )
-        
-        // Assert
-        #expect(response.rulesSummary.title == "Monopoly Rules")
-        #expect(response.processedGameTitle == "Monopoly")
-        #expect(response.generatedAt <= Date())
-        #expect(response.wasCached == false)
-        
-        // Test cached response
-        let cachedResponse = GenerateRulesUseCase.Response(
-            rulesSummary: rulesSummary,
-            processedGameTitle: "Monopoly",
-            wasCached: true
-        )
-        #expect(cachedResponse.wasCached == true)
-    }
-    
-    /// Test use case dependency injection pattern.
-    @Test("GenerateRulesUseCase follows dependency injection patterns")
-    func testDependencyInjectionPattern() async throws {
-        await world.resetAll()
-        // This test validates that the GenerateRulesUseCase follows the established
-        // dependency injection pattern used throughout the application
-        
-        // The use case requires these dependencies to be injected via constructor:
-        // - rulesOrchestrationService: RulesOrchestrationService
-        // - aiInputValidator: AIInputValidatorServiceInterface
-        // - cacheKeyGenerator: CacheKeyGeneratorServiceInterface
-        // - aiCache: AICacheServiceInterface
-        // - llmService: LLMService
-        // - aiResponseValidator: AIResponseValidationService
-        // - cacheConfiguration: CacheConfig
-        
-        // This validates the architectural decision for use case dependency injection,
-        // which differs from domain services that inject dependencies at method level
-        
-        // Use cases encapsulate complete business workflows and need stable dependencies
-        // Domain services coordinate multiple external services and need flexible injection
-        
-        // For testing, we would create the use case with mock dependencies
-        // This test validates the pattern exists and is followed correctly
-        
-        let context = RequestContext(
-            clientIP: "127.0.0.1",
-            logger: Logger(label: "test")
-        )
-        
-        let request = GenerateRulesUseCase.Request(
-            gameTitle: "Test Game",
-            context: context
-        )
-        
-        #expect(request.gameTitle == "Test Game")
-        #expect(request.context.clientIP == "127.0.0.1")
-    }
-    
-    /// Test Command protocol compliance.
-    @Test("GenerateRulesUseCase implements Command protocol correctly")
-    func testCommandProtocolCompliance() async throws {
-        await world.resetAll()
-        // The GenerateRulesUseCase should implement the Command protocol
-        // which defines the interface for all use cases in the system
-        
-        // Command protocol ensures:
-        // 1. Consistent interface across all use cases
-        // 2. Clear separation of request/response patterns
-        // 3. Testable business logic encapsulation
-        // 4. Clean integration with controllers
-        
-        // This test validates the architectural pattern rather than implementation
-        // since testing the actual command would require complex mock setup
-        
-        let context = RequestContext(
-            clientIP: "192.168.1.1",
-            logger: Logger(label: "command-test")
-        )
-        
-        let request = GenerateRulesUseCase.Request(
-            gameTitle: "Command Pattern Game",
-            context: context
-        )
-        
-        // Validate request structure matches Command pattern expectations
-        #expect(request.gameTitle == "Command Pattern Game")
-        #expect(request.context.clientIP == "192.168.1.1")
-    }
-    
-    /// Test response JSON serialization for API endpoints.
-    @Test("GenerateRulesUseCase Response can be serialized for API responses")
-    func testResponseSerialization() async throws {
-        await world.resetAll()
-        // Arrange
-        let rulesSummary = RulesSummary.Response(
-            title: "Scythe Rules",
-            playerCount: "1-5",
-            playTime: "90-115 minutes",
-            summary: "An engine-building game set in an alternate-history 1920s",
-            initialSetup: ["Choose faction", "Place workers"],
-            firstRoundGuide: ["Take actions", "Gain resources"],
-            winCondition: "Control territories and complete objectives",
-            deepDive: ["Combat strategy", "Resource management"],
-            resources: RulesSummary.Response.GameResources(
-                videoLinks: [],
-                webLinks: []
-            ),
-            confidence: 92,
-            notes: "Based on official rulebook"
-        )
-        
-        let response = GenerateRulesUseCase.Response(
-            rulesSummary: rulesSummary,
-            processedGameTitle: "Scythe",
-            generatedAt: Date(),
-            wasCached: true
-        )
-        
-        // Act - Test that nested structure can be encoded
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .iso8601
-        
-        let jsonData = try encoder.encode(response.rulesSummary)
-        
-        // Assert - JSON encoding succeeds
-        #expect(jsonData.count > 0)
-        
-        // Decode to verify round-trip integrity
-        let decoder = JSONDecoder()
-        let decodedSummary = try decoder.decode(RulesSummary.Response.self, from: jsonData)
-        
-        #expect(decodedSummary.title == rulesSummary.title)
-        #expect(decodedSummary.playerCount == rulesSummary.playerCount)
-        #expect(decodedSummary.confidence == rulesSummary.confidence)
-    }
-    
-    /// Test error handling patterns for use case failures.
-    @Test("GenerateRulesUseCase handles expected error types")
-    func testErrorHandlingPatterns() async throws {
-        await world.resetAll()
-        // The use case should handle these error scenarios:
-        // 1. Invalid game title (ValidationError)
-        // 2. AI service failures (ContentError)
-        // 3. Cache failures (graceful degradation)
-        // 4. Security validation failures (AIValidationError)
-        
-        // Test validation errors
-        let emptyTitleError = AIProcessingError.emptyInput(context: "game_title")
-        #expect(emptyTitleError.description.contains("game_title"))
-        
-        let emptyInputError = AIProcessingError.emptyInput(context: "generic_input")
-        #expect(emptyInputError.description.contains("generic_input"))
-        
-        // Test AI validation errors  
-        let emptyDataError = AIProcessingError.imageDataEmpty
-        #expect(emptyDataError.description.contains("empty"))
-        
-        // Error types are properly defined and can be thrown/caught
-        // The use case would catch these errors and transform them into
-        // appropriate HTTP responses via the controller layer
-    }
-
-    /// Ensure persisted records are returned when cache misses.
-    @Test("GenerateRulesUseCase rehydrates from persisted summary on cache miss")
-    func testReturnsPersistedSummaryWhenAvailable() async throws {
-        await world.resetAll()
-        cache.configureForceMiss(true)
-
-        let sanitizedTitle = try app.serviceCache.aiInputValidatorService
-            .validateAndSanitizeGameTitle("Catan: Persisted")
-        let cacheKey = app.serviceCache.cacheKeyGeneratorService.generateRulesKey(for: sanitizedTitle)
-        let createdAt = Date(timeIntervalSince1970: 5_000)
-
-        let persisted = GeneratedRuleModel(
-            originalTitle: "Catan: Persisted",
-            sanitizedTitle: sanitizedTitle,
-            cacheKey: cacheKey,
-            title: "Persisted Catan",
-            playerCount: "3-4",
-            playTime: "60-90 minutes",
-            summary: "Use stored summary when available",
-            initialSetup: ["Build the island"],
-            firstRoundGuide: ["Roll dice", "Collect resources"],
-            winCondition: "Reach 10 points",
-            deepDive: ["Trade smart"],
-            resourcesVideoLinks: ["https://videos.example/persisted"],
-            resourcesWebLinks: ["https://persisted.example"],
-            confidence: 88,
-            notes: "Persisted from integration test",
-            lastAccessedAt: Date(timeIntervalSince1970: 0)
-        )
-        persisted.createdAt = createdAt
-        try await repository.create(persisted)
-
-        let useCase = try await app.serviceRegistry.resolveRequired(GenerateRulesUseCase.self)
-        let request = GenerateRulesUseCase.Request(
-            gameTitle: "Catan: Persisted",
-            context: RequestContext(clientIP: "127.0.0.1", logger: app.logger)
-        )
-
-        let response = try await useCase.execute(request)
-
-        #expect(response.rulesSummary.title == "Persisted Catan")
-        #expect(response.wasCached == false)
-        #expect(response.generatedAt == createdAt)
-
-        cache.configureForceMiss(false)
-        let hydrated = await cache.get(key: cacheKey)
-        #expect(hydrated != nil)
-
-        let touched = try await repository.find(bySanitizedTitle: sanitizedTitle)
-        #expect(touched?.lastAccessedAt ?? Date(timeIntervalSince1970: 0) > Date(timeIntervalSince1970: 0))
-    }
-
-    /// Ensure new summaries are persisted and cached after LLM generation.
-    @Test("GenerateRulesUseCase persists new summaries after generation")
-    func testPersistsSummaryAfterGeneration() async throws {
-        await world.resetAll()
-        cache.configureForceMiss(true)
-
-        let sanitizedTitle = try app.serviceCache.aiInputValidatorService
-            .validateAndSanitizeGameTitle("Terraforming Mars")
-        let cacheKey = app.serviceCache.cacheKeyGeneratorService.generateRulesKey(for: sanitizedTitle)
-
-        let llmResponse = """
-        {
-          "title": "Terraforming Mars Deluxe",
-          "playerCount": "1-5",
-          "playTime": "120 minutes",
-          "summary": "Guide corporations to terraform the red planet.",
-          "initialSetup": ["Select corporations", "Deal project cards"],
-          "firstRoundGuide": ["Choose actions", "Manage heat production"],
-          "winCondition": "Highest terraforming rating wins",
-          "deepDive": ["Focus on engine building", "Balance oxygen and oceans"],
-          "resources": {
-            "videoLinks": ["https://videos.example/tm"],
-            "webLinks": ["https://terraformingmars.example"]
-          },
-          "confidence": 82,
-          "notes": "Assumes Prelude expansion"
-        }
-        """
-        world.llm.setDefaultResponse(llmResponse)
-
-        let useCase = try await app.serviceRegistry.resolveRequired(GenerateRulesUseCase.self)
-        let request = GenerateRulesUseCase.Request(
-            gameTitle: "Terraforming Mars",
-            context: RequestContext(clientIP: "192.168.0.1", logger: app.logger)
-        )
-
-        let response = try await useCase.execute(request)
-
-        #expect(response.rulesSummary.title == "Terraforming Mars Deluxe")
-        #expect(response.wasCached == false)
-
-        let stored = try await repository.find(bySanitizedTitle: sanitizedTitle)
-        #expect(stored != nil)
-        #expect(stored?.summary == "Guide corporations to terraform the red planet.")
-
-        cache.configureForceMiss(false)
-        let cached = await cache.get(key: cacheKey)
-        #expect(cached != nil)
-    }
-}
-
-// MARK: - Complex Command Testing Pattern Note
-
-/*
-This test demonstrates Complex Command (Use Case) testing patterns:
-
-1. **Request/Response Structure Testing**: Validating data contracts and interfaces
-2. **Dependency Injection Validation**: Ensuring architectural patterns are followed  
-3. **Protocol Compliance**: Testing adherence to Command pattern interfaces
-4. **JSON Serialization**: Testing API response compatibility and data integrity
-5. **Error Handling Patterns**: Validating expected error types and behaviors
-6. **Business Logic Interfaces**: Testing use case contracts without implementation
-
-Key characteristics of complex command testing:
-- Focus on business logic interfaces and data contracts
-- Test architectural patterns and dependency management
-- Validate complete request/response cycles for API compatibility
-- Ensure error handling follows established patterns
-- Test data integrity through serialization boundaries
-- Validate business logic encapsulation without mocking complexity
-
-This approach provides:
-- Stable tests that survive implementation changes
-- Clear validation of business logic interfaces
-- Fast execution without complex service mocking
-- Focus on use case contracts and data flow
-- Easy maintenance as business requirements evolve
-- Comprehensive validation of architectural decisions
-
-Complex commands differ from simple commands by:
-- Orchestrating multiple domain services
-- Managing complex multi-step business workflows
-- Handling sophisticated caching and validation strategies
-- Providing high-level business capability interfaces
-- Coordinating cross-cutting concerns like security and performance
-- Encapsulating complete user stories and business operations
-
-Testing philosophy for commands:
-- Test the interface, not the implementation
-- Validate business logic contracts and data flow
-- Ensure architectural patterns are properly implemented
-- Focus on use case capabilities and error handling
-- Maintain separation between business logic and infrastructure
-- Provide clear validation of business requirements fulfillment
-*/
